Deletado os arquivos desnecessarios da classe controlador. Mudado o tipo do identificador das câmeras para inteiro sem sinal de 64 bits.

+++ silvver_cameras/src/abstractCamera.hpp
@@ -34,7 +34,7 @@ protected:
 
   CameraConfig cameraConfig;
 
-  unsigned UID;
+  unsigned long UID;
 
   bool stopping;
